+++ spooq/transformer/null_cleaner.py
@@ -122,7 +122,7 @@ class NullCleaner(BaseCleaner):
                     F.col(column_name).isNull(),
                     substitute,
                 ).otherwise(F.col(column_name))
-                .cast(data_type)
+                .try_cast(data_type)
             )
             self.logger.debug("Cleansing Expression for " + column_name + ": " + str(cleansing_expression))
             cleansing_expressions.append((column_name, cleansing_expression))