+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+4.0.0 (2025-12-01) - BREAKING CHANGES! Only Spark 4+!
+--------------------------------------
+* [MOD] Spark 4+ support! Spark `<4` is no longer supported. Please use Spooq version 3.4.2 for Spark 2.x and 3.x versions.
+* [MOD] Mapper: Remove support for string parameters (`mode` and `missing_column_handling`).
+* [MOD] Mapper: Remove support for transformations defined as string (`"extended_string_to_double"` vs `to_double()`).
+
 3.4.2 (2024-08-08)
 -------------------
 * [ADD] Annotator: New transformer to load and apply comments to dataframes