+++ docs/source/changelog.rst
@@ -1,1 +1,1 @@
-../../CHANGELOG.rst
+CHANGELOG.rst